Update simple_rdbms.py
Altered the relevant database methods to check if the table it's referring to exists before performing any actions on it.  Also added a View class to represent views, and initialized two dictionaries ('views' and 'queries') in the Database class to provide a place to store queries and views separately from the set of main tables.  Finally, a dropTable() method was added to give the option to drop a table from the database.