Update simple_rdbms.py
In the previous revision(s), any rows added by the insert method to one table are added to the other tables.  This  was rectified by converting the inner table to a list, append the row, and then re-convert the list into a set.  Another change was changing the getTable() method to the getInnerTable() method in the Table class.  Also simplified the cartesian_product() function for the time being.