Update simple_rdbms.py
Added the ability of the database to create a table with attributes and their corresponding data types (ex. int, str, date).  This was done by modifying attr in the Table class and insert() method.  

Also added the displayTableHeadings() method in the Database class to allow users to display the attributes and their data types in the selected table.