Update simple_rdbms.py
Added methods: getTotalSize() in both Database and SimpleRDBMS classes, and getTableSize() in Database class.  These methods give a more reasonable picture of the size of the database system in terms of its contents. 

Also added methods: getNumRows() in Table class and getTableRows() in Database class.  These allow the system to provide the user with the number of rows a certain table has, and it displays this information when the user is inserting and (hopefully) deleting in tables.